+++ src/main/java/com/boxesapp/desktopapp/model/dao/Dao.java
@@ -6,6 +6,6 @@ public interface Dao <T> {
     public void create(T t);
     public void update(T t);
     public void delete(T t);
-    public T findById(Long id);
+    public T findById(Integer id);
     public List<T> findAll();
 }
